feat(SPC_ONEDXINTERNAL-8888): Tính năng thi ở client

+++ QuizIT/QuizIT.Service/Entities/History.cs
@@ -13,7 +13,7 @@ namespace QuizIT.Service.Entities
         public int Id { get; set; }
         public int ExamId { get; set; }
         public int UserId { get; set; }
-        public int TimeDoExam { get; set; }
+        public double TimeDoExam { get; set; }
         public int Point { get; set; }
         public DateTime? CreatedAt { get; set; }
